requirements.txt
llama-index
llama-index-readers-file
openai
python-dotenv